+++ src/main/java/fullydynamic/FullyDynamicExhaustiveCountingThreeNode.java
@@ -18,14 +18,14 @@ public class FullyDynamicExhaustiveCountingThreeNode implements TopkGraphPattern
 	NodeMap nodeMap;
 	EdgeHandler utility;
 	THashMap<Triplet, Integer> counter;
-	THashMap<Pattern, Long> frequentPatterns;
+	THashMap<Pattern, Integer> frequentPatterns;
 	int numSubgraph;
 	public FullyDynamicExhaustiveCountingThreeNode() {
 		this.nodeMap = new NodeMap();
 		utility = new EdgeHandler();
 		counter = new THashMap<>();
 		numSubgraph = 0 ;
-		frequentPatterns = new THashMap<Pattern, Long>();
+		frequentPatterns = new THashMap<Pattern, Integer>();
 	}
 	@Override
 	public boolean addEdge(StreamEdge edge) {
@@ -162,17 +162,17 @@ public class FullyDynamicExhaustiveCountingThreeNode implements TopkGraphPattern
 	void addFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count+1);
 		}else {
-			frequentPatterns.put(p, 1l);
+			frequentPatterns.put(p, 1);
 		}
 	}
 	
 	void removeFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			if(count >1)
 				frequentPatterns.put(p, count-1);
 			else 
@@ -181,7 +181,7 @@ public class FullyDynamicExhaustiveCountingThreeNode implements TopkGraphPattern
 	}
 	
 	@Override
-	public THashMap<Pattern, Long> getFrequentPatterns() {
+	public THashMap<Pattern, Integer> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 	@Override
@@ -193,7 +193,7 @@ public class FullyDynamicExhaustiveCountingThreeNode implements TopkGraphPattern
 		return 0;
 	}
 	@Override
-	public THashMap<Pattern, Long> correctEstimates() {
+	public THashMap<Pattern, Integer> correctEstimates() {
 		return frequentPatterns;
 	}
