+++ src/main/java/fullydynamic/FullyDynamicSubgraphReservoirFourNode.java
@@ -25,7 +25,7 @@ public class FullyDynamicSubgraphReservoirFourNode implements TopkGraphPatterns 
 	NodeMap nodeMap;
 	EdgeHandler utility;
 	AdvancedSubgraphReservoir<Quadriplet> reservoir;
-	THashMap<Pattern, Long> frequentPatterns;
+	THashMap<Pattern, Integer> frequentPatterns;
 	int N; // total number of subgraphs
 	int M; // maximum reservoir size
 	int Ncurrent;
@@ -51,7 +51,7 @@ public class FullyDynamicSubgraphReservoirFourNode implements TopkGraphPatterns 
 		c2=0;
 		sum = 0;
 		Zprime=-1;
-		frequentPatterns = new THashMap<Pattern, Long>();
+		frequentPatterns = new THashMap<Pattern, Integer>();
 		skipRS = new AlgorithmZ(M);
 		//skipRP = new AlgorithmD();
 		sampler = new ReservoirSampling<LabeledNode>();
@@ -271,17 +271,17 @@ public class FullyDynamicSubgraphReservoirFourNode implements TopkGraphPatterns 
 	void addFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count+1);
 		}else {
-			frequentPatterns.put(p, 1l);
+			frequentPatterns.put(p, 1);
 		}
 	}
 
 	void removeFrequentPattern(Quadriplet t) {
 		FourNodeGraphPattern p = new FourNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			if(count >1)
 				frequentPatterns.put(p, count-1);
 			else 
@@ -290,18 +290,18 @@ public class FullyDynamicSubgraphReservoirFourNode implements TopkGraphPatterns 
 	}
 
 	@Override
-	public THashMap<Pattern, Long> getFrequentPatterns() {
+	public THashMap<Pattern, Integer> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 	@Override
-	public THashMap<Pattern, Long> correctEstimates() {
-		THashMap<Pattern, Long> correctFrequentPatterns = new THashMap<>();
+	public THashMap<Pattern, Integer> correctEstimates() {
+		THashMap<Pattern, Integer> correctFrequentPatterns = new THashMap<>();
 		double correctFactor = correctFactor();
 		List<Pattern> patterns = new ArrayList<Pattern>(frequentPatterns.keySet());
 		for(Pattern p: patterns) {
 			long count = frequentPatterns.get(p);
 			double value = count*correctFactor;
-			correctFrequentPatterns.put(p, (long)value);
+			correctFrequentPatterns.put(p, (int)value);
 		}
 		return correctFrequentPatterns;
 	}
