+++ src/main/java/fullydynamic/FullyDynamicSubgraphReservoirThreeNode4.java
@@ -24,7 +24,7 @@ public class FullyDynamicSubgraphReservoirThreeNode4 implements TopkGraphPattern
 	NodeMap nodeMap;
 	EdgeHandler utility;
 	SubgraphReservoir<Triplet> reservoir;
-	THashMap<Pattern, Long> frequentPatterns;
+	THashMap<Pattern, Integer> frequentPatterns;
 	int N; // total number of subgraphs
 	int M; // maximum reservoir size
 	int Ncurrent;
@@ -46,7 +46,7 @@ public class FullyDynamicSubgraphReservoirThreeNode4 implements TopkGraphPattern
 		c2=0;
 		sum = 0;
 		Zprime=-1;
-		frequentPatterns = new THashMap<Pattern, Long>();
+		frequentPatterns = new THashMap<Pattern, Integer>();
 		skipRS = new AlgorithmZ(M);
 		skipRP = new AlgorithmD();
 		sampler = new ReservoirSampling<Triplet>();
@@ -249,17 +249,17 @@ public class FullyDynamicSubgraphReservoirThreeNode4 implements TopkGraphPattern
 	void addFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			frequentPatterns.put(p, count+1);
 		}else {
-			frequentPatterns.put(p, 1l);
+			frequentPatterns.put(p, 1);
 		}
 	}
 
 	void removeFrequentPattern(Triplet t) {
 		ThreeNodeGraphPattern p = new ThreeNodeGraphPattern(t);
 		if(frequentPatterns.containsKey(p)) {
-			long count = frequentPatterns.get(p);
+			int count = frequentPatterns.get(p);
 			if(count >1)
 				frequentPatterns.put(p, count-1);
 			else 
@@ -268,18 +268,18 @@ public class FullyDynamicSubgraphReservoirThreeNode4 implements TopkGraphPattern
 	}
 
 	@Override
-	public THashMap<Pattern, Long> getFrequentPatterns() {
+	public THashMap<Pattern, Integer> getFrequentPatterns() {
 		return this.frequentPatterns;
 	}
 	@Override
-	public THashMap<Pattern, Long> correctEstimates() {
-		THashMap<Pattern, Long> correctFrequentPatterns = new THashMap<Pattern, Long>();
+	public THashMap<Pattern, Integer> correctEstimates() {
+		THashMap<Pattern, Integer> correctFrequentPatterns = new THashMap<Pattern, Integer>();
 		double correctFactor = correctFactor();
 		List<Pattern> patterns = new ArrayList<Pattern>(frequentPatterns.keySet());
 		for(Pattern p: patterns) {
 			long count = frequentPatterns.get(p);
 			double value = count*correctFactor;
-			correctFrequentPatterns.put(p, (long)value);
+			correctFrequentPatterns.put(p, (int)value);
 		}
 		return correctFrequentPatterns;
 	}
